# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation",
        "description": "L5 Swagger OpenApi description",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://employeestracker.loc/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/transactions": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "get list of transactions",
                "description": "Agar search parametri bosa qidiradi, bomasa transactions larni qaytaradi ",
                "operationId": "GetTransactions",
                "parameters": [
                    {
                        "name": "search",
                        "in": "path",
                        "description": "Userlarni qididrish first_name yoki last_name orqali",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Transactions were taken successfully"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "create a new transaction",
                "description": "create a new transaction",
                "operationId": "StoreTransactions",
                "requestBody": {
                    "description": "create a new transaction",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type",
                                    "where",
                                    "dateTime"
                                ],
                                "properties": {
                                    "type": {
                                        "type": "integer",
                                        "example": "4"
                                    },
                                    "where": {
                                        "type": "integer",
                                        "example": "2"
                                    },
                                    "dateTime": {
                                        "type": "dateTime",
                                        "example": "2022-01-27 11:29:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Transaction is saved successfully!!!"
                                        },
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/transactions/{id}/edit": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "get a transaction data",
                "description": "get a transaction data",
                "operationId": "EditTransactions",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Kerak bo'lgan transaction id si",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Transaction was taken successfully"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/transactions/{id}": {
            "put": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Returns updated transaction data",
                "description": "Returns updated transaction data",
                "operationId": "updateTransaction",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "transaction id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "update the transaction",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type",
                                    "where",
                                    "dateTime"
                                ],
                                "properties": {
                                    "type": {
                                        "type": "integer",
                                        "example": "4"
                                    },
                                    "where": {
                                        "type": "integer",
                                        "example": "2"
                                    },
                                    "dateTime": {
                                        "type": "dateTime",
                                        "example": "2022-01-27 11:29:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "correct response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Transaction is updated successfully!!!"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Delete existing transaction",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteTransaction",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Transaction id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Transaction is deleted successfully!!!"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/transactions/user-transactions": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "description": "get list of User transactions",
                "operationId": "GetUserTransactions",
                "parameters": [
                    {
                        "name": "from",
                        "in": "path",
                        "description": "DateTime qachondan boshlashi",
                        "required": true,
                        "schema": {
                            "type": "date"
                        }
                    },
                    {
                        "name": "to",
                        "in": "path",
                        "description": "DateTime qachongacha filtrlash",
                        "required": true,
                        "schema": {
                            "type": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User transactions were taken successfully"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Something wrong"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/transactions/user-last-transactions": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "get list of User Last transactions",
                "description": "get list of User Last transactions",
                "operationId": "GetUserLastTransactions",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Your last transaction is taken successfully"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "There is not this user transaction"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    }
}